+++ app/build.gradle
@@ -23,7 +23,6 @@
  */
 
 apply plugin: 'com.android.application'
-apply plugin: 'kotlin-platform-android'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
