+++ app/src/main/java/com/perol/asdpl/pixivez/manager/DownloadManagerActivity.kt
@@ -3,9 +3,6 @@ package com.perol.asdpl.pixivez.manager
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.view.Menu
-import android.view.MenuItem
-import com.arialyy.aria.core.Aria
 import com.perol.asdpl.pixivez.R
 import com.perol.asdpl.pixivez.activity.RinkActivity
 import com.perol.asdpl.pixivez.objects.ThemeUtil
@@ -13,29 +10,8 @@ import kotlinx.android.synthetic.main.activity_download_manager.*
 
 
 class DownloadManagerActivity : RinkActivity() {
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        menuInflater.inflate(R.menu.menu_download_manager, menu)
-        return true
 
-    }
 
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-
-        when (item.itemId) {
-            R.id.action_settings -> {
-                startActivity(Intent(this, ManagerSettingsActivity::class.java))
-            }
-            R.id.action_resume -> {
-                Aria.download(this).resumeAllTask()
-            }
-            R.id.action_cancel -> {
-                Aria.download(this).removeAllTask(false);
-            }
-
-        }
-        return super.onOptionsItemSelected(item)
-
-    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
